+++ lib/commands/customers-inactive.d.ts
@@ -1,0 +1,9 @@
+import { CommandEvent, Result } from "../types";
+/**
+ * Event To Sync Inactive Quickbooks Customers - Repzo Clients
+ * @param commandEvent
+ * @returns
+ */
+export declare const inactive_customers: (
+  commandEvent: CommandEvent
+) => Promise<Result>;