+++ src/commands/index.ts
@@ -8,6 +8,9 @@ import { join } from "./join.js";
 import { basic } from "./basic.js";
 import { oAuth2 } from "./oAuth2.js";
 import { preferences } from "./preferences.js";
+import { inventory_adjustment } from "./inventory-adjustment.js";
+import { inactive_items } from "./items-inactive.js";
+import { inactive_customers } from "./customers-inactive.js";
 
 /**
  * Route Command Event - Function
@@ -33,6 +36,12 @@ export const commands = async (CommandEvent: CommandEvent) => {
         return await join(CommandEvent);
       case "oAuth2":
         return await oAuth2(CommandEvent);
+      case "inventory_adjustment":
+        return await inventory_adjustment(CommandEvent);
+      case "inactive_items":
+        return await inactive_items(CommandEvent);
+      case "inactive_customers":
+        return await inactive_customers(CommandEvent);
       default:
         throw `Route: ${CommandEvent.command} not found`;
     }
@@ -64,15 +73,33 @@ export const commandsList: Command[] = [
       "this command to sync all quickbooks customers with your repzo clients",
   },
   {
-    command: "sync_products",
-    name: "Sync products",
+    command: "inactive_customers",
+    name: "Sync Inactive Clients",
     description:
-      "this command to syncing all quickbooks items with your repzo products",
+      "this command to sync all inactive quickbooks customers with your repzo clients",
   },
   {
     command: "sync_taxs",
     name: "Sync taxes",
     description:
       "this command to syncing all quickbooks items with your repzo taxes",
+  },
+  {
+    command: "sync_products",
+    name: "Sync products",
+    description:
+      "this command to syncing all quickbooks items with your repzo products",
+  },
+  {
+    command: "inactive_items",
+    name: "Sync Inactive products",
+    description:
+      "this command to syncing all inactive quickbooks items with your repzo products",
+  },
+  {
+    command: "inventory_adjustment",
+    name: "Inventory Adjustment",
+    description:
+      "this command shall create an inventory adjustment with the difference between all accumulated repzo warehouses and quickbooks inventory making quickbooks similar to Repzo accumulated inventory",
   },
 ];