+++ src/quickbooks/types/Item.ts
@@ -77,11 +77,11 @@ export namespace Item {
   interface itemBody {
     Name: string;
     Type: string;
-    QtyOnHand: number;
-    IncomeAccountRef: ReferenceType;
-    AssetAccountRef: ReferenceType;
-    ExpenseAccountRef: ReferenceType;
-    InvStartDate: Date;
+    QtyOnHand?: number;
+    IncomeAccountRef?: ReferenceType;
+    AssetAccountRef?: ReferenceType;
+    ExpenseAccountRef?: ReferenceType;
+    InvStartDate?: Date | string;
   }
 
   export namespace Find {